backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Anthropic Claude
anthropic==0.18.1

# OpenAI (for Whisper STT and TTS)
openai==1.12.0

# Audio processing
pydub==0.25.1

# Image processing
Pillow==10.2.0

# Environment variables
python-dotenv==1.0.1

# Session management
uuid==1.30

# Async support
aiofiles==23.2.1
aiohttp==3.9.3

# CORS
starlette==0.35.1
